shared token middlewareを削除
