+++ apps/kaede/src/middleware/request-actor.test.ts
@@ -89,6 +89,25 @@ describe('requestActorMiddleware', () => {
     expect(findValidSessionByTokenMock).not.toHaveBeenCalled()
   })
 
+  it('Bearer scheme の大小文字と余分な空白を許容する', async () => {
+    const app = createTestApp()
+
+    const response = await app.request('/api/ping', {
+      headers: {
+        authorization: '  bearer    shared-token  ',
+      },
+    })
+
+    expect(response.status).toBe(200)
+    await expect(response.json()).resolves.toEqual({
+      actor: {
+        type: 'service_client',
+        name: 'shared_token',
+      },
+    })
+    expect(findValidSessionByTokenMock).not.toHaveBeenCalled()
+  })
+
   it('不正な Bearer token は session cookie fallback せず拒否する', async () => {
     mockActiveSessionUser()
     const app = createTestApp()
